move toTableData to grafana/ui
